added like/dislike/remove video buttons. They all work. Want to add an alert when the user removes a video to make sure they really want.
- TODO :
- alert(); when the user clicks the remove video button to make sure they really want to.
- animation for the rating buttons / an animation for removing the video
-  reformat React code to make it cleaner / move into other components
- IMPORTANT :
- add new video button that opens a form on click where the user can submit a new video
- desktop breakpoints
